🔨 [services] Move constants `UPDATE_BRANCH` and `LATEST_BRANCH` to `create`
