--- src/cutty/services/git.py
@@ -1,3 +1,0 @@
-"""Storing files in a Git repository."""
-LATEST_BRANCH = "cutty/latest"
-UPDATE_BRANCH = "cutty/update"
