+++ src/cutty/services/update.py
@@ -7,8 +7,8 @@ from typing import Optional
 
 from cutty.repositories.domain.repository import Repository as Template
 from cutty.services.create import create
-from cutty.services.git import LATEST_BRANCH
-from cutty.services.git import UPDATE_BRANCH
+from cutty.services.create import LATEST_BRANCH
+from cutty.services.create import UPDATE_BRANCH
 from cutty.templates.adapters.cookiecutter.projectconfig import readprojectconfigfile
 from cutty.templates.domain.bindings import Binding
 from cutty.util.git import Repository
