+++ tests/unit/services/test_git.py
@@ -12,8 +12,8 @@ from cutty.filestorage.domain.observers import observe
 from cutty.filestorage.domain.storage import FileStorage
 from cutty.filesystems.domain.purepath import PurePath
 from cutty.services.create import creategitrepository
-from cutty.services.git import LATEST_BRANCH
-from cutty.services.git import UPDATE_BRANCH
+from cutty.services.create import LATEST_BRANCH
+from cutty.services.create import UPDATE_BRANCH
 from cutty.util.git import Repository
 from tests.util.git import createbranches
