+++ tests/unit/services/test_update.py
@@ -7,8 +7,8 @@ import pytest
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path as VirtualPath
 from cutty.repositories.domain.repository import Repository as Template
-from cutty.services.git import LATEST_BRANCH
-from cutty.services.git import UPDATE_BRANCH
+from cutty.services.create import LATEST_BRANCH
+from cutty.services.create import UPDATE_BRANCH
 from cutty.services.update import abortupdate
 from cutty.services.update import continueupdate
 from cutty.services.update import CreateProject
